requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
black==21.10b0
celery==5.2.7
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.1.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-framework-datatables-filter==0.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-extra-fields==3.2.1
drf-writable-nested==0.6.3
drf-yasg==1.21.3
flake8==4.0.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.3
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.9
python-decouple==3.5
pytz==2021.3
redis==4.3.4
regex==2022.9.13
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.14.1
